refactor(daemon): rename daemon.js → manager.js source files (Phase 1)
Phase 1 of designs/daemon-rename-to-manager.md: mechanical file renames
only, no identifier renames. Aligns the JS orchestration layer's file
names with the Rust `endor` supervisor, which already calls this role the
manager (ManagerMode, ENDO_MANAGER_NODE, spawn_inproc_xs_manager).

`git mv` per the design's File-renames table (13 files):
  daemon.js, daemon-{node,node-powers,go,go-powers,database,
  database-node,persistence-powers,webextension}.js, and
  bus-daemon-{node,node-powers,rust-xs,rust-xs-powers}.js
  → the corresponding manager-*/bus-manager-* names.

Only import specifiers and runtime path references that point at the
renamed files are updated:
  - relative `import`/`@import`/`import()` specifiers across the package
  - the `@endo/daemon/src/daemon-node-powers.js` deep-subpath export key
    in package.json (→ manager-node-powers.js) and its consumers
    (agent-tools, agentry test fixtures)
  - runtime `new URL('src/…')` / `path.join(…, 'src/…')` references in
    index.js, familiar/scripts/bundle.mjs,
    daemon/scripts/bundle-bus-daemon-rust-xs.mjs, familiar/resource-paths.js
  - two self-referential startup error strings inside the renamed files

`Daemon*`/`Mignonic` identifiers, the `@endo/daemon` package name, the
`packages/daemon/` directory, `bench-daemon.js`, and
`bus-xs-daemon-polyfills.js` are intentionally unchanged (Phase 2/kept).

Verified: `tsc` clean; `ava test/endo.test.js` 200/200 pass.
