+++ packages/agent-tools/test/git-remote-fixtures.js
@@ -9,7 +9,7 @@ import { promisify } from 'node:util';
 import { makeNativeGitBackend } from '@endo/git';
 import { makeGit } from '@endo/exo-git';
 import { makeMount, lineageOf } from '@endo/daemon/src/mount.js';
-import { makeFilePowers } from '@endo/daemon/src/daemon-node-powers.js';
+import { makeFilePowers } from '@endo/daemon/src/manager-node-powers.js';
 
 const execFileAsync = promisify(execFile);
