+++ packages/daemon/package.json
@@ -32,7 +32,7 @@
     "./type-guards.js": "./src/type-guards.js",
     "./src/interfaces.js": "./src/interfaces.js",
     "./src/mount.js": "./src/mount.js",
-    "./src/daemon-node-powers.js": "./src/daemon-node-powers.js",
+    "./src/manager-node-powers.js": "./src/manager-node-powers.js",
     "./package.json": "./package.json"
   },
   "scripts": {
