+++ packages/daemon/src/bus-manager-node.js
@@ -2,24 +2,24 @@
 /* global process */
 
 // Establish a perimeter:
-// eslint-disable-next-line import/order
 import '@endo/init';
 
 import crypto from 'crypto';
 import net from 'net';
 import fs from 'fs';
+import fsp from 'fs/promises';
 import path from 'path';
 import url from 'url';
 
 import { E } from '@endo/far';
 import { makePromiseKit } from '@endo/promise-kit';
-import { makeDaemon } from './daemon.js';
+import { makeDaemon } from './manager.js';
 import {
   makeFilePowers,
   makeNetworkPowers,
   makeCryptoPowers,
-} from './daemon-node-powers.js';
-import { makeDaemonicGoPowers } from './daemon-go-powers.js';
+} from './manager-node-powers.js';
+import { makeDaemonicBusPowers } from './bus-manager-node-powers.js';
 import {
   encodeEnvelope,
   decodeEnvelope,
@@ -27,24 +27,20 @@ import {
   writeFrameToStream,
 } from './envelope.js';
 
-const fsp = { access: fs.promises.access };
-
 /** @import { PromiseKit } from '@endo/promise-kit' */
 /** @import { Config } from './types.js' */
 
-// The daemon-go entry point receives its configuration as command-line
-// arguments, identical to daemon-node.js. It also opens fd 3/4 for the
-// envelope protocol with the engo supervisor.
+// Bus daemon entry point — the Node.js daemon process running under an
+// external supervisor (Go or Rust). Receives configuration as command-line
+// arguments and communicates with the supervisor via the envelope protocol
+// on fd 3/4.
 //
-// fd 3: daemon writes envelopes to engo (child → parent)
-// fd 4: daemon reads envelopes from engo (parent → child)
-//
-// Note: This matches the Go subprocess convention where fd 3 is the
-// child's write end and fd 4 is the child's read end.
+// fd 3: daemon writes envelopes to supervisor (child → parent)
+// fd 4: daemon reads envelopes from supervisor (parent → child)
 
 if (process.argv.length < 5) {
   throw new Error(
-    `daemon-go.js requires arguments [sockPath] [statePath] [ephemeralStatePath] [cachePath], got ${process.argv.join(
+    `bus-manager-node.js requires arguments [sockPath] [statePath] [ephemeralStatePath] [cachePath], got ${process.argv.join(
       ', ',
     )}`,
   );
@@ -69,21 +65,21 @@ const envelopeReadStream = fs.createReadStream(null, { fd: 4 });
 // @ts-ignore This is in fact how you open a file descriptor.
 const envelopeWriteStream = fs.createWriteStream(null, { fd: 3 });
 
-// Read the init envelope from engo to learn our handle.
+// Read the init envelope from the supervisor to learn our handle.
 const readInitEnvelope = async () => {
   const frameData = await readFrameFromStream(envelopeReadStream);
   if (frameData === null) {
-    throw new Error('daemon-go: EOF before init envelope');
+    throw new Error('bus-daemon: EOF before init envelope');
   }
   const env = decodeEnvelope(frameData);
   if (env.verb !== 'init') {
-    throw new Error(`daemon-go: expected init envelope, got ${env.verb}`);
+    throw new Error(`bus-daemon: expected init envelope, got ${env.verb}`);
   }
   return env.handle;
 };
 
 /**
- * Send an envelope to engo.
+ * Send an envelope to the supervisor.
  * @param {number} handle
  * @param {string} verb
  * @param {Uint8Array} [payload]
@@ -99,15 +95,11 @@ const sendEnvelope = async (handle, verb, payload, nonce) => {
   await writeFrameToStream(envelopeWriteStream, data);
 };
 
-const { promise: cancelled, reject: cancel } =
-  /** @type {PromiseKit<never>} */ (makePromiseKit());
-
 const networkPowers = makeNetworkPowers({ net, fsp });
 const filePowers = makeFilePowers({ fs, path });
 const cryptoPowers = makeCryptoPowers(crypto);
-const powers = await makeDaemonicGoPowers({
+const powers = makeDaemonicBusPowers({
   config,
-  cancelled,
   url,
   filePowers,
   cryptoPowers,
@@ -116,10 +108,13 @@ const powers = await makeDaemonicGoPowers({
 });
 const { persistence: daemonicPersistencePowers } = powers;
 
-// Engo owns endo.pid (the authoritative PID for kill). This function
-// overwrites it with the node daemon PID so that killDaemonProcess
+const { promise: cancelled, reject: cancel } =
+  /** @type {PromiseKit<never>} */ (makePromiseKit());
+
+// The supervisor owns endo.pid (the authoritative PID for kill). This
+// function overwrites it with the node daemon PID so that killDaemonProcess
 // targets the process that owns the socket and its child workers.
-// Killing the node daemon causes engo to detect the exit and cascade.
+// Killing the node daemon causes the supervisor to detect the exit and cascade.
 const updateRecordedPid = async () => {
   const pidPath = filePowers.joinPath(ephemeralStatePath, 'endo.pid');
   await filePowers.writeFileText(pidPath, `${pid}\n`);
@@ -127,36 +122,34 @@ const updateRecordedPid = async () => {
 
 const main = async () => {
   const daemonLabel = `daemon on PID ${pid}`;
-  console.log(`Endo daemon (go platform) starting on PID ${pid}`);
+  console.log(`Endo daemon (bus) starting on PID ${pid}`);
   cancelled.catch(() => {
-    console.log(`Endo daemon (go platform) stopping on PID ${pid}`);
+    console.log(`Endo daemon (bus) stopping on PID ${pid}`);
   });
 
-  // Read init envelope from engo supervisor.
+  // Read init envelope from supervisor.
   const selfHandle = await readInitEnvelope();
-  console.log(`Endo daemon assigned handle ${selfHandle} by engo`);
+  console.log(`Endo daemon assigned handle ${selfHandle} by supervisor`);
 
   // Start the envelope reader now that the init envelope has been consumed.
   // This must happen after readInitEnvelope() to avoid a race on fd 4.
-  powers.control.startEnvelopeReader && powers.control.startEnvelopeReader();
+  if (powers.control.startEnvelopeReader) {
+    powers.control.startEnvelopeReader();
+  }
 
   await daemonicPersistencePowers.initializePersistence();
 
-  const {
-    endoBootstrap,
-    cancelGracePeriod,
-    capTpConnectionRegistrar,
-    marshalSaveError,
-  } = await makeDaemon(
-    powers,
-    daemonLabel,
-    cancel,
-    cancelled,
-    {},
-    {
-      defaultWorkerKind: 'locked',
-    },
-  );
+  const { endoBootstrap, cancelGracePeriod, capTpConnectionRegistrar } =
+    await makeDaemon(
+      powers,
+      daemonLabel,
+      cancel,
+      cancelled,
+      {},
+      {
+        defaultWorkerKind: 'locked',
+      },
+    );
 
   /** @param {Error} error */
   const exitWithError = error => {
@@ -171,7 +164,6 @@ const main = async () => {
     cancelled,
     exitWithError,
     capTpConnectionRegistrar,
-    marshalSaveError,
   );
   const services = [privatePathService];
 
@@ -184,9 +176,9 @@ const main = async () => {
     const agentIdPath = filePowers.joinPath(statePath, 'root');
     await filePowers.writeFileText(agentIdPath, `${agentId}\n`);
 
-    // Signal readiness to engo supervisor.
+    // Signal readiness to supervisor.
     await sendEnvelope(0, 'ready');
-    console.log('Endo daemon (go platform) ready, signaled engo');
+    console.log('Endo daemon (bus) ready, signaled supervisor');
   } catch (error) {
     // No IPC to parent like daemon-node.js; just log and throw.
     console.error('Daemon startup failed:', error);
@@ -196,7 +188,7 @@ const main = async () => {
   const servicesStopped = Promise.all(services.map(({ stopped }) => stopped));
 
   // Record self as official daemon process so killDaemonProcess targets
-  // the node daemon (which owns workers) rather than engo.
+  // the node daemon (which owns workers) rather than the supervisor.
   await updateRecordedPid();
 
   // Wait for services to end normally
