+++ packages/daemon/src/manager-go-powers.js
@@ -1,43 +1,31 @@
-// @ts-nocheck
-// The bus Node daemon powers still delegate to pre-SQLite pet-store
-// and persistence factories; aligning with the llm-side constructor
-// signatures (single daemonDb argument) is follow-up work.
-
+// @ts-check
+/* eslint-disable no-await-in-loop, no-bitwise */
 /* global process */
-/* eslint-disable no-void */
-/* eslint-disable no-bitwise -- CBOR envelope encoding / length-prefix
-   parsing requires bitwise operators. */
-/* eslint-disable no-continue -- Envelope framing loop skips malformed
-   frames without nesting. */
-/* eslint-disable no-await-in-loop -- Envelope reader awaits each
-   frame serially by design. */
 
 /**
- * Daemonic powers for the bus platform (external supervisor).
+ * Daemonic powers for the Go (engo) platform.
  *
- * Instead of spawning workers via child_process.fork(), this module
- * requests worker spawns from the supervisor via the envelope protocol.
+ * This module is a derivative of daemon-node-powers.js. It shares all powers
+ * except for control powers (makeWorker), which requests worker spawns from
+ * the engo supervisor via the envelope protocol instead of using
+ * child_process.fork().
  *
  * The envelope protocol runs on fd 3 (child→parent) and fd 4 (parent→child),
  * using CBOR-framed envelopes.
  */
 
 import { makePromiseKit } from '@endo/promise-kit';
-import { mapWriter, mapReader } from '@endo/stream';
 import { bytesFromText } from '@endo/bytes/from-string.js';
 import { bytesToText } from '@endo/bytes/to-string.js';
-import {
-  makeMessageCapTP,
-  messageToBytes,
-  bytesToMessage,
-} from './connection.js';
+import { makeNetstringCapTP } from './connection.js';
 import { makePetStoreMaker } from './pet-store.js';
 import { decodeEnvelope, readFrameFromStream } from './envelope.js';
 import {
   makeFilePowers,
   makeCryptoPowers,
   makeDaemonicPersistencePowers,
-} from './daemon-node-powers.js';
+} from './manager-node-powers.js';
+import { makeDaemonDatabase } from './manager-database-node.js';
 
 /** @import { ERef } from '@endo/eventual-send' */
 /** @import { CapTpConnectionRegistrar, Config, CryptoPowers, DaemonWorkerFacet, DaemonicPowers, FilePowers, WorkerDaemonFacet } from './types.js' */
@@ -52,15 +40,14 @@ export { makeFilePowers, makeCryptoPowers, makeDaemonicPersistencePowers };
  */
 
 /**
- * Create control powers that spawn workers through the bus supervisor.
+ * Create control powers that spawn workers through the engo supervisor.
  *
  * Instead of calling child_process.fork(), this sends a "spawn" envelope
- * to the supervisor (handle 0). The supervisor spawns the worker as a peer
- * subprocess and routes messages between daemon and worker via handle
- * rewriting.
+ * to engo (handle 0). Engo spawns the worker as a peer subprocess and
+ * routes messages between daemon and worker via handle rewriting.
  *
  * CapTP traffic between daemon and worker is encapsulated in envelopes
- * with verb "deliver". The daemon demultiplexes incoming envelopes by
+ * with verb "captp". The daemon demultiplexes incoming envelopes by
  * handle to dispatch to the correct per-worker CapTP session.
  *
  * @param {Config} config
@@ -68,7 +55,7 @@ export { makeFilePowers, makeCryptoPowers, makeDaemonicPersistencePowers };
  * @param {FilePowers} filePowers
  * @param {EnvelopeChannel} envelopeChannel
  */
-export const makeDaemonicBusControlPowers = (
+export const makeDaemonicGoControlPowers = (
   config,
   fileURLToPath,
   filePowers,
@@ -76,14 +63,9 @@ export const makeDaemonicBusControlPowers = (
 ) => {
   const { sendEnvelope, envelopeReadStream } = envelopeChannel;
 
-  const endoWorkerBin = process.env.ENDO_WORKER_BIN || '';
-  const endoNodeWorkerBin = process.env.ENDO_NODE_WORKER_BIN || '';
   const endoWorkerPath =
     process.env.ENDO_WORKER_SUBPROCESS_PATH ||
-    fileURLToPath(new URL('bus-worker-node.js', import.meta.url));
-  const endoNodeWorkerPath = fileURLToPath(
-    new URL('bus-worker-node-raw.js', import.meta.url),
-  );
+    fileURLToPath(new URL('worker-go.js', import.meta.url));
 
   /** @type {Map<number, { resolve: (env: import('./envelope.js').Envelope) => void }>} */
   const pendingSpawns = new Map();
@@ -98,7 +80,7 @@ export const makeDaemonicBusControlPowers = (
 
   // Start the envelope reader loop. This demultiplexes incoming envelopes:
   // - Envelopes with verb "spawned" are responses to spawn requests.
-  // - Envelopes with verb "deliver" are CapTP messages from workers,
+  // - Envelopes with verb "captp" are CapTP frames from workers,
   //   dispatched to the per-worker CapTP session.
   const startEnvelopeReader = () => {
     const readLoop = async () => {
@@ -117,10 +99,7 @@ export const makeDaemonicBusControlPowers = (
             pendingSpawns.delete(env.nonce);
             pending.resolve(env);
           }
-          continue;
-        }
-
-        if (env.verb === 'error' && env.nonce > 0) {
+        } else if (env.verb === 'error' && env.nonce > 0) {
           // Error response to a spawn request.
           const pending = pendingSpawns.get(env.nonce);
           if (pending) {
@@ -134,28 +113,18 @@ export const makeDaemonicBusControlPowers = (
               }),
             );
           }
-          continue;
-        }
-
-        if (env.verb === 'deliver') {
-          // CapTP message from a worker. The handle field has been
-          // rewritten by the supervisor to the worker's handle.
+        } else if (env.verb === 'captp') {
+          // CapTP frame from a worker. The handle field has been
+          // rewritten by engo to the worker's handle.
           const workerHandle = env.handle;
           const entry = workerWriters.get(workerHandle);
           if (entry) {
-            // Deliver the CapTP message to the per-worker reader.
-            // The payload is the raw JSON-encoded CapTP message.
+            // Deliver the CapTP frame to the per-worker reader.
+            // The payload is the raw netstring-framed CapTP bytes.
             void entry.writer.next(env.payload);
-          } else {
-            console.error(
-              `bus-daemon: deliver from unknown worker handle=${workerHandle}`,
-            );
           }
-          continue;
-        }
-
-        if (env.verb === 'exited') {
-          // Worker exit notification from the supervisor.
+        } else if (env.verb === 'exited') {
+          // Worker exit notification from engo.
           const workerHandle = env.handle;
           const resolve = workerExitResolvers.get(workerHandle);
           if (resolve) {
@@ -167,22 +136,21 @@ export const makeDaemonicBusControlPowers = (
           if (entry) {
             void entry.writer.return(undefined);
           }
-          continue;
+        } else {
+          // Unhandled envelope verb — log and continue.
+          console.error(
+            `daemon-go: unhandled envelope verb=${env.verb} handle=${env.handle}`,
+          );
         }
-
-        // Unhandled envelope verb — log and continue.
-        console.error(
-          `bus-daemon: unhandled envelope verb=${env.verb} handle=${env.handle}`,
-        );
       }
     };
     void readLoop().catch(error => {
-      console.error('bus-daemon: envelope reader error:', error);
+      console.error('daemon-go: envelope reader error:', error);
     });
   };
 
   /**
-   * Spawn a worker through the bus supervisor.
+   * Spawn a worker through engo.
    *
    * @param {string} workerId
    * @param {DaemonWorkerFacet} daemonWorkerFacet
@@ -191,7 +159,11 @@ export const makeDaemonicBusControlPowers = (
    * @param {CapTpConnectionRegistrar} [capTpConnectionRegistrar]
    * @param {string[]} [_trustedShims]
    * @param {string} [_label]
-   * @param {'locked' | 'node'} [kind]
+   * @param {'locked' | 'node'} [_kind]
+   *   Worker kind. Currently unused by the Go powers implementation,
+   *   but accepted to keep the positional arity aligned with the type
+   *   in `types.d.ts` so `marshalLoadError` lands in the correct slot.
+   * @param {(err: Error, errorId?: string) => void} [marshalLoadError]
    */
   const makeWorker = async (
     workerId,
@@ -199,9 +171,10 @@ export const makeDaemonicBusControlPowers = (
     cancelled,
     _forceCancelled,
     capTpConnectionRegistrar = undefined,
-    _trustedShims = undefined,
-    _label = undefined,
-    kind = 'locked',
+    _trustedShims = undefined, // eslint-disable-line no-underscore-dangle
+    _label = undefined, // eslint-disable-line no-underscore-dangle
+    _kind = undefined, // eslint-disable-line no-underscore-dangle
+    marshalLoadError = undefined,
   ) => {
     const { statePath, ephemeralStatePath } = config;
 
@@ -217,21 +190,12 @@ export const makeDaemonicBusControlPowers = (
       filePowers.makePath(workerEphemeralStatePath),
     ]);
 
-    // Build spawn request. For kind === 'node', use ENDO_NODE_WORKER_BIN
-    // (or fall back to bus-worker-node-raw.js). Otherwise use
-    // ENDO_WORKER_BIN (or fall back to bus-worker-node.js).
-    let workerParts;
-    if (kind === 'node') {
-      workerParts = endoNodeWorkerBin
-        ? endoNodeWorkerBin.split(/\s+/).filter(Boolean)
-        : [process.execPath, endoNodeWorkerPath];
-    } else {
-      workerParts = endoWorkerBin
-        ? endoWorkerBin.split(/\s+/).filter(Boolean)
-        : [process.execPath, endoWorkerPath];
-    }
-    const command = workerParts[0];
-    const args = workerParts.slice(1);
+    // Find the node executable.
+    const nodePath = process.execPath;
+
+    // Build spawn request: engo will spawn `node worker-go.js`
+    const command = nodePath;
+    const args = [endoWorkerPath];
 
     // Allocate a nonce for the spawn request.
     const nonce = nextNonce;
@@ -247,12 +211,12 @@ export const makeDaemonicBusControlPowers = (
     // Encode the spawn request payload as CBOR map {command, args}.
     const payloadBuf = encodeSpawnPayload(command, args);
 
-    // Send spawn request to supervisor (handle 0).
+    // Send spawn request to engo (handle 0).
     await sendEnvelope(0, 'spawn', payloadBuf, nonce);
 
     console.log(`Endo worker spawn requested for ${workerId} (nonce=${nonce})`);
 
-    // Wait for the supervisor to respond with the worker's handle.
+    // Wait for engo to respond with the worker's handle.
     const response = await spawnResponse;
 
     if (response.verb === 'error') {
@@ -268,8 +232,8 @@ export const makeDaemonicBusControlPowers = (
     // Set up CapTP session over the envelope protocol.
     // Create a pipe pair for the CapTP layer. The daemon side writes
     // CapTP frames that get wrapped in envelopes and sent to the worker
-    // via the supervisor. Incoming CapTP frames from the worker arrive via
-    // the envelope reader and are pushed into the reader side of the pipe.
+    // via engo. Incoming CapTP frames from the worker arrive via the
+    // envelope reader and are pushed into the reader side of the pipe.
     const { makePipe } = await import('@endo/stream');
     const [captpReadFrom, captpWriteTo] = makePipe();
 
@@ -277,13 +241,12 @@ export const makeDaemonicBusControlPowers = (
     // CapTP frames.
     workerWriters.set(workerHandle, { writer: captpWriteTo });
 
-    // Create a writer that wraps CapTP messages in envelopes addressed
-    // to the worker's handle. The payload is the JSON-encoded message
-    // bytes (no netstring framing — the envelope protocol provides framing).
+    // Create a writer that wraps CapTP frames in envelopes addressed
+    // to the worker's handle.
     /** @type {import('@endo/stream').Writer<Uint8Array>} */
-    const envelopeBytesWriter = harden({
+    const envelopeCapTPWriter = harden({
       async next(/** @type {Uint8Array} */ chunk) {
-        await sendEnvelope(workerHandle, 'deliver', chunk);
+        await sendEnvelope(workerHandle, 'captp', chunk);
         return harden({ done: false, value: undefined });
       },
       async return(/** @type {undefined} */ _value) {
@@ -297,22 +260,19 @@ export const makeDaemonicBusControlPowers = (
       },
     });
 
-    const messageWriter = mapWriter(envelopeBytesWriter, messageToBytes);
-    const messageReader = mapReader(captpReadFrom, bytesToMessage);
-
     const workerClosed = new Promise(resolve => {
       workerExitResolvers.set(workerHandle, resolve);
       // Also resolve on cancellation so shutdown isn't blocked.
       cancelled.catch(() => resolve(undefined));
     });
 
-    const { getBootstrap, closed: capTpClosed } = makeMessageCapTP(
+    const { getBootstrap, closed: capTpClosed } = makeNetstringCapTP(
       `Worker ${workerId}`,
-      messageWriter,
-      messageReader,
+      envelopeCapTPWriter,
+      captpReadFrom,
       cancelled,
       daemonWorkerFacet,
-      undefined,
+      { marshalLoadError },
       capTpConnectionRegistrar,
     );
 
@@ -340,15 +300,17 @@ export const makeDaemonicBusControlPowers = (
 /**
  * @param {object} opts
  * @param {Config} opts.config
+ * @param {Promise<never>} opts.cancelled
  * @param {typeof import('url')} opts.url
  * @param {FilePowers} opts.filePowers
  * @param {CryptoPowers} opts.cryptoPowers
  * @param {(handle: number, verb: string, payload?: Uint8Array, nonce?: number) => Promise<void>} opts.sendEnvelope
  * @param {import('stream').Readable} opts.envelopeReadStream
- * @returns {DaemonicPowers}
+ * @returns {Promise<DaemonicPowers>}
  */
-export const makeDaemonicBusPowers = ({
+export const makeDaemonicGoPowers = async ({
   config,
+  cancelled,
   url,
   filePowers,
   cryptoPowers,
@@ -357,13 +319,20 @@ export const makeDaemonicBusPowers = ({
 }) => {
   const { fileURLToPath } = url;
 
-  const petStorePowers = makePetStoreMaker(filePowers, config);
+  // Ensure state directory exists before opening database.
+  await filePowers.makePath(config.statePath);
+
+  const daemonDb = makeDaemonDatabase(config);
+  cancelled.catch(() => daemonDb.close());
+
+  const petStorePowers = makePetStoreMaker(daemonDb);
   const daemonicPersistencePowers = makeDaemonicPersistencePowers(
+    daemonDb,
     filePowers,
     cryptoPowers,
     config,
   );
-  const daemonicControlPowers = makeDaemonicBusControlPowers(
+  const daemonicControlPowers = makeDaemonicGoControlPowers(
     config,
     fileURLToPath,
     filePowers,
@@ -414,7 +383,7 @@ const cborHead = (buf, major, n) => {
 
 /**
  * Encode a spawn request payload as CBOR: {command: text, args: [text...]}.
- * This matches the supervisor's DecodeSpawnRequest format.
+ * This matches the Go DecodeSpawnRequest format.
  * @param {string} command
  * @param {string[]} args
  * @returns {Uint8Array}
