+++ packages/daemon/src/manager.js
