+++ packages/daemon/test/git-remote.test.js
@@ -25,7 +25,7 @@ import {
   revokeGitCredential,
 } from '@endo/exo-git';
 
-import { makeFilePowers } from '../src/daemon-node-powers.js';
+import { makeFilePowers } from '../src/manager-node-powers.js';
 import { lineageOf, makeMount } from '../src/mount.js';
 
 const execFileAsync = nodePromisify(execFile);
