+++ packages/daemon/test/mount-platform-fs-conformance.test.js
@@ -21,8 +21,8 @@ import {
   checkinTree,
 } from '@endo/platform/fs/lite';
 
-import { makeFilePowers } from '../src/daemon-node-powers.js';
-import { makeXsFilePowers } from '../src/bus-daemon-rust-xs-powers.js';
+import { makeFilePowers } from '../src/manager-node-powers.js';
+import { makeXsFilePowers } from '../src/bus-manager-rust-xs-powers.js';
 import { makeMount } from '../src/mount.js';
 import { makeMemoryStore } from './_mount-test-helpers.js';
