+++ packages/familiar/src/resource-paths.js
@@ -50,7 +50,7 @@ if (isPackaged) {
   nodePath = 'node';
   endoCliPath = path.join(repoRoot, 'packages/cli/bin/endo.cjs');
   chatDistPath = path.join(repoRoot, 'packages/chat/dist/index.html');
-  endoDaemonPath = path.join(repoRoot, 'packages/daemon/src/daemon-node.js');
+  endoDaemonPath = path.join(repoRoot, 'packages/daemon/src/manager-node.js');
   workerSubprocessPath = path.join(
     repoRoot,
     'packages/daemon/src/worker-node.js',
